fix(test): replace "BLACK_HOLE_ADDRESS" with real one
IPv6 has a preserved address which acts as a black hole (see [RFC 6666](https://tools.ietf.org/html/rfc6666) )

Questions: does this break on agents which do not support IPv6?

Another option for a IPv4 black hole might be from `240.0.0.0/4`, which is "reserved for future use"  (e.g. see https://superuser.com/questions/698244/ip-address-that-is-the-equivalent-of-dev-null )

Signed-off-by: Maximilian Huber <external.maximilian.huber2@bosch-si.com>