chores(quick-deploy): add quick deploy for portlet
This commit is an addition to the quick deploy PR sw360/sw360-chores#55

In order to prevent the Tomcat from creating a temp directory for the webapp
from which the webapp is service, the following context settings are
necessary:
  - antiJARLocking=false
  - antiResourceLocking=false

If the webapp provides no context file, liferay generates such a file with
both settings set to true.

This commit introduces an own context file which is filled based on properties.
These propterties are "true" by default but can be set to "false" by using
the maven profile "develop" on building.

Signed-off-by: Leo von Klenze <leo.vonklenze@scansation.de>