feat(licenses): add support for project-only obligations
This change introduces a new portlet for managing TODOs under the Admin
menu. There are 2 views. One for listing TODOs saved in the database and
one for adding a new one. In the listing view, the delete action is also
available but can only be executed by users with SW360_ADMIN rights and
above.

This new PrivatePages.lar includes the TODOs management portlet.

Signed-off-by: Nikolas Sepos <nikolas@endocode.com>