chores(developer): remove developer tag in pom.xml
During lifetime of SW360 many developers have contributed to the
code. This contribution was highly appreciated and is still honored.

However, the developer tag inside the pom.xml file in each submodule
was not maintained most of the time. Therefore we decided to remove
this outdated information.

This is in line with the rules of the Maven project for this tag:
only developers who should be contacted in case of questions should
be mentioned there (https://maven.apache.org/pom.html#Developers).

Aware that all contributors are still mentioned through their
helpful commits in the git history, this commits removes the
outdated tags.