chores(deploy): add new deploy profile
With Liferay 7.x there is no deploy mechanism from Liferay anymore.
In addtion each submodule must be copied to a different location
depending on its type:
- backend:    backend submodules must be deployed directly into the Tomcat.
              Deploying to the liferay auto-deploy-folder does no longer work.
- frontend:   these submodules must be deployed to the auto-deploy folder of
              Liferay.
- libraries:  these submodules are packed as OSGi modules and they are
              dependencies of the frontend modules. Therefore they must be
              deployed to the Liferay auto-deploy foldre as well.

This is also an improvement for setups that are using a different
application server for the Backend and Liferay.

See the README.md file for details about the deployment.

Signed-off-by: Leo von Klenze <leo.vonklenze@scansation.de>