chores(build): Fix deploy profile
Before this commit, the maven dependency plugin was activated with
the 'deploy' profile for every module, even for the parent pom. Since
there is a configuration to copy some information from the
'build-configuration'-submodule, an initial build with active profile
failed.

This fix moves the configuration of the plugin for the 'deploy'-profile
into a 'pluginManagement'-section. This way it is only activated for
relevant modules.

This also removes the old profile sections.

Closes #621

Signed-off-by: Leo von Klenze <leo.vonklenze@scansation.de>