feature(table-filter): add filter box, fix print
This commits adds an option to the generic create-method for datatables as
well as fixes the print button on some pages.

* Filter Box
There is no one more parameter to the create-method of the bridges/datatables
module. Giving true as last parameter will add a filter box at the top right
for the datatable.
This option has been set on many pages now, except where explicit quick filters
already exist.

* Fix print button
When adding a print button, the additional option 'orthogonal' has been set.
With this option in place, all render methods are called with type set to
'print' when printing. This way renderer can omit html tags or links when
rendering for print view. Renderes has been adapted as necessary.

Closes #657

Signed-off-by: Leo von Klenze <leo.vonklenze@scansation.de>