feature(reelase): add release merge wizard
This commit adds a merge wizard for releases. It is build upon the merge framework therefore
it works the same as the component and the vendor merge wizard: click merge on the target, select
a source and takeover the needed fields.

Merging releases is only possible inside a single component. In addition both releases need to have
at least on pair of equal source attachemnts nor no source attachments at all.

The fossology status is not merges since it will be superseeded by external tool requests in the
near future.

Signed-off-by: Leo von Klenze <leo.vonklenze@scansation.de>