Add tests for hashes endpoint and documentation in asciidoc
Add tests in ReleaseSpecTest to test getting and adding hash
information of a release.
Add a "knownHash" to the release in the test file and add a
responsefield to the respondFields of updating and getting a
release.
While developing a NPE was discovered, when the getOtherKnownHashes
of a release was null, and fixed in ReleaseController.

Signed-off-by: Stephanie Neubauer <Stephanie.Neubauer@bosch-si.com>