chore(script): Added support to uninstall the current thrift version.
The script to install thrift has been extended to accept a new
"--uninstall" parameter. If this parameter is passed, it checks
whether thrift is currently installed on the local machine. If so, the
thrift version is determined, and this version is uninstalled.

This is useful when the thrift version is updated. Then the old
version needs to be uninstalled first before the new one can be
installed.

When calling the script to install thrift, but an unsupported version
is found, it now prints a message referring to the new parameter.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>