fix(rest): Add dirty context to resource health indicator test
Issue: #896

Since per SpringBootTest class the context gets loaded only
once, the result of previous test somehow seem to modify the
application context. This solution was used:
https://stackoverflow.com/a/13288671/6466630

It is not a problem of the cache. Introducing a
NoOpCacheManager which should simply take cache items
without storing them, did not solve the problem.

Signed-off-by: Stephanie Neubauer <Stephanie.Neubauer@bosch.io>