bug(eclipse): Quickfix for maven flatten pom problem
The mechanism with the ci profile in the main pom.xml does not work, because the double
setting of patchlevel is used correctly to determine the version for maven build
but the wrong version is used by the flatten pom plugin. The fix ignores
the ci profile and uses only one patchlevel variable. The price is that
the minor version of SW360 is now managed in the Jenkinsfile not the pom.xml.

Signed-off-by: Lars Geyer-Blaumeiser <lars.geyer-blaumeiser@bosch.io>