fix(build): Fix issue with overwriting of patchlevel variable in pom.xml
Issue is that the maven flatten module is not able to handle the
double definition of variable patchlevel in the pom correctly.
The generated pom files have the wrong version.
Solution: Put the minor version in the revision variable and
reduce the standard patchlevel content to the idea from the "ci"
profile. Remove the "ci" profile and adapt the travis and jenkins
build files.

Removed the git count mechanism from travis build script because
it was only decorating the naming of build steps but had no
impact on build execution

Signed-off-by: Lars Geyer-Blaumeiser <lars.geyer-blaumeiser@bosch.io>