fix(deps): Fixed wrong dependency download
Guava dependency was pointing out to wrong directory and not providing
any error on output, so creating an invalid file that failis only on
unpacking level.
If this kept used, in future maven should do the download automatically
instead of a script.

This provided possibility to alter Dockerfile deps download position,
and avoid recreate the build layer all the time.

In addition, a dockerignore file is added to avoid copy deploy / target
/ buiid massive outputs if local build is used to build docker too.

Signed-off-by: Helio Chissini de Castro <heliocastro@gmail.com>