upd(thrift): Use only provided tarball to generate resources
Thrift need to be built inside docker, but was only compiler, not
required jar, which is downloaded as external dependency.

Now both compiler and jar are generated on the fly from same build, and
using cmake based build. Boost and ssl dev dependencies are no more
necessary to the builder base image, reducing size and install time for
base build system.

Signed-off-by: Helio Chissini de Castro <heliocastro@gmail.com>