fix(docker): Remove deletion that invalidate image
A logic flaw due to a lack of information was also introduced when the cleanup
packages workflow was introduced.

Images on ghcr do not contain the layers info in their API, which leads to being
unable to restrict the deletion of nonchanged layers for a long time.
This is especially valid for base, clucene, and thrift layers.

The unattended removal of old layers makes the main image (sw360) not reachable with
the error of manifest unknown.

Reference issue: https://github.com/eclipse-sw360/sw360/issues/2240

Signed-off-by: Helio Chissini de Castro <heliocastro@gmail.com>