feat(backend): use spring DI containers
Use Spring DI Containers for the Attachment Backend servlet. This allows
solving the constructor overload issue.

The class `AbstractBackendServletInitializer` provides the basics of
methods which allows the Servlets to initialize themselves. This class
also loads the properties from `couchdb.properties` or
`couchdb-test.properties` based on the active profile for beans
creation.

Signed-off-by: Gaurav Mishra <mishra.gaurav@siemens.com>