fix(test): fix testcases for attachments
Migrate the test cases for attachments backend to spring based test
cases. Note that the test case was initially creating the DB and
initializing the handler afterwards. But with Spring, the handler is
initialized while creating the containers and thus the creation of DB in
`setUp()` removes the views from Repository. Thus, the test case now
uses `@DirtiesContext` to create new objects at the end of each test
case and drops the DB in `tearDown()` to provide fresh DB for next test
case.

Signed-off-by: Gaurav Mishra <mishra.gaurav@siemens.com>